join-service/requirements.txt
flask
opencv-python
requests
pika
redis
google-cloud-storage

